requirements.txt
Django==5.1.4
django-filter==25.1
django-rest-swagger==2.2.0
drf-yasg==1.21.7
gunicorn==23.0.0
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1